+++ components/device-mgt/org.wso2.carbon.device.mgt.core/src/gen/java/org/wso2/carbon/device/mgt/core/api/AdminApi.java
@@ -17,40 +17,30 @@
  *
  */
 
-package org.wso2.carbon.device.mgt.core;
+package org.wso2.carbon.device.mgt.core.api;
 
 import io.swagger.annotations.ApiParam;
 
-import org.wso2.carbon.device.mgt.core.dto.DeviceType;
-import org.wso2.carbon.device.mgt.core.dto.ErrorResponse;
-import org.wso2.carbon.device.mgt.core.factories.AdminApiServiceFactory;
+import org.wso2.carbon.device.mgt.core.api.dto.DeviceType;
+import org.wso2.carbon.device.mgt.core.api.factories.AdminApiServiceFactory;
 
 import org.wso2.msf4j.Microservice;
 import org.wso2.msf4j.Request;
-import org.wso2.msf4j.formparam.FileInfo;
-import org.wso2.msf4j.formparam.FormDataParam;
 import org.osgi.service.component.annotations.Component;
 
-import java.io.InputStream;
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
-import javax.ws.rs.HEAD;
 
-import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 
 @Component(
-        name = "org.wso2.carbon.device.mgt.core.AdminApi",
+        name = "AdminApi",
         service = Microservice.class,
         immediate = true
 )
