+++ components/device-mgt/org.wso2.carbon.device.mgt.core/src/gen/java/org/wso2/carbon/device/mgt/core/api/ApiOriginFilter.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.wso2.carbon.device.mgt.core;
+package org.wso2.carbon.device.mgt.core.api;
 
 import java.io.IOException;
