+++ components/device-mgt/org.wso2.carbon.device.mgt.core/src/gen/java/org/wso2/carbon/device/mgt/core/api/Application.java
@@ -1,4 +1,4 @@
-package org.wso2.carbon.device.mgt.core;
+package org.wso2.carbon.device.mgt.core.api;
 
 import org.wso2.msf4j.MicroservicesRunner;
