+++ components/device-mgt/org.wso2.carbon.device.mgt.core/src/gen/java/org/wso2/carbon/device/mgt/core/api/dto/MetadataEntry.java
@@ -1,9 +1,7 @@
-package org.wso2.carbon.device.mgt.core.dto;
+package org.wso2.carbon.device.mgt.core.api.dto;
 
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 
 import java.util.Objects;
