+++ components/device-mgt/org.wso2.carbon.device.mgt.core/src/main/java/org/wso2/carbon/device/mgt/core/api/factories/AdminApiServiceFactory.java
@@ -17,13 +17,15 @@
  *
  */
 
-package org.wso2.carbon.device.mgt.core;
+package org.wso2.carbon.device.mgt.core.api.factories;
 
-public class NotFoundException extends ApiException {
-    private int code;
+import org.wso2.carbon.device.mgt.core.api.AdminApiService;
+import org.wso2.carbon.device.mgt.core.api.impl.AdminApiServiceImpl;
 
-    public NotFoundException(int code, String msg) {
-        super(code, msg);
-        this.code = code;
+public class AdminApiServiceFactory {
+    private static final AdminApiService service = new AdminApiServiceImpl();
+
+    public static AdminApiService getAdminApi() {
+        return service;
     }
 }
